fixed graphics and set locale + using long doble for better persicion (but it didn't help)

+++ src/controllers/s21_credit_controller.cc
@@ -1,7 +1,8 @@
 #include "s21_credit_controller.h"
 
 namespace s21 {
-void CreditController::SetStartValues(double loan, double period, double rate) {
+void CreditController::SetStartValues(long double loan, long double period,
+                                      long double rate) {
   credit_calc.SetStartValues(loan, period, rate);
 }
 
@@ -11,23 +12,23 @@ void CreditController::CalcDifferentiated() {
   credit_calc.CalcDifferentiated();
 }
 
-double CreditController::GetMonthPayment() {
+long double CreditController::GetMonthPayment() {
   return credit_calc.GetMonthPayment();
 }
 
-double CreditController::GetOverpayment() {
+long double CreditController::GetOverpayment() {
   return credit_calc.GetOverpayment();
 }
 
-double CreditController::GetTotalPayment() {
+long double CreditController::GetTotalPayment() {
   return credit_calc.GetTotalPayment();
 }
 
-double CreditController::GetFirstPayment() {
+long double CreditController::GetFirstPayment() {
   return credit_calc.GetFirstPayment();
 }
 
-double CreditController::GetLastPayment() {
+long double CreditController::GetLastPayment() {
   return credit_calc.GetLastPayment();
 }
 }  // namespace s21
