+++ src/models/s21_credit_calc.h
@@ -9,28 +9,28 @@
 class CreditCalculator {
  public:
   CreditCalculator() = default;
-  CreditCalculator(double loan, int period, double rate)
+  CreditCalculator(long double loan, int period, long double rate)
       : loan_(loan), period_(period), rate_(rate){};
   ~CreditCalculator() = default;
 
-  void SetStartValues(double loan, int period, double rate);
+  void SetStartValues(long double loan, int period, long double rate);
   void CalcAnnuity();
   void CalcDifferentiated();
-  double GetMonthPayment() { return month_payment_; }
-  double GetTotalPayment() { return total_payment_; }
-  double GetOverpayment() { return overpayment_; }
-  double GetFirstPayment() { return first_payment_; }
-  double GetLastPayment() { return last_payment_; }
+  long double GetMonthPayment() { return month_payment_; }
+  long double GetTotalPayment() { return total_payment_; }
+  long double GetOverpayment() { return overpayment_; }
+  long double GetFirstPayment() { return first_payment_; }
+  long double GetLastPayment() { return last_payment_; }
 
  private:
-  double CalcMonthAnnuity();
-  double loan_{};
+  long double CalcMonthAnnuity();
+  long double loan_{};
   int period_{};
-  double rate_{};
-  double first_payment_{}, last_payment_{};
-  double month_payment_{}, total_payment_{}, overpayment_{};
+  long double rate_{};
+  long double first_payment_{}, last_payment_{};
+  long double month_payment_{}, total_payment_{}, overpayment_{};
   static constexpr int days_in_year_ = 365;
-  static constexpr double months_in_year_ = 12.0;
+  static constexpr long double months_in_year_ = 12.0;
 };
 
 #endif  // SMARTCALC_CPP_SRC_MODEL_CREDITCALCULATOR_H
